Streamline top level section markup
The "CHANGELOG.rst" and "README.rst" files used differing markup
characters for the top level section headings. As these files are
concatenated to generate the long description on the packages PyPI
landing page, this lead to a "Title level inconsistent:" error when
running `tox -e check`.
One alternative way of fixing that would be to drop one level from all
section headings in "CHANGELOG.rst", but if we think that this would
look better, we could also try it later.