Configure `AIRFLOW_HOME` in `egon-data`
We want airflow to work with the DAGs, configuration files and other
stuff that gets shipped with this package, so we have to point the
`AIRFLOW_HOME` environment variable dynamically. In order for this to
work, the corresponding module, i.e. `egon.data.airflow` has to be
made importable via an "__init__.py" file, module's `__file__` attribute
exists and we can get at the containing directory.