Forward `egon-data airflow` to `airflow`
This is the bare minimum of stuff that should work: calling
`egon-data airflow` should simply call `airflow` with `AIRFLOW_HOME` set
to the directory shipped with this package and with all options
forwarded. The `context_settings` line enables argument forwarding
via `context.args` while `add_help_option=False` makes sure that
`--help` is correctly forwarded to `airflow` and is not intercepted by
`click`. This is ensured by the test.