Synchronize isort's configuration with black's style
Since I recommend using [`black`][0] to format code, making [`isort`][1]
format imports in a way that is compatible with `black` is kind of
important. Fortunately `black` has a nice guide for this and since
version five, `isort` already has a built-in profile respecting this.
The only thing that we have to manually adjust is the line length, which
we set to PEP8 compatible 79 characters. And of course we also have to
skip one of the repository template files, because that one uses a
different style convention.
Since this configuration change has direct implications on code, the
affected imports have of course been adjusted. Also, since profiles are
only available as of `isort` version 5, this constraint has been added
in to "tox.ini".

[0]: https://black.readthedocs.io/en/stable/index.html
[1]: https://pycqa.github.io/isort/