Blacken "setup.py" and "src/egon/data/cli.py"
As always they're run through `black -l 79`. Apparently, multi line
docstrings are not allowed to start with a space. Also, apparently
`black` doesn't want the `entry_points` dictionary to be on a single
line, although I could've sworn that `black` did this by itself...
Should stay the way it is now, though. :)