Add a `pre-commit` hook for `black`
Done as [documented][0]. Since one has to have `black` installed in
order to make use of this hook, I added `black` to the dev dependencies.
The "pyproject.toml" file is necessary, because that's the only place
that `black` will look for configuration parameters. Last but not least,
the `[build-system]` section in "pyproject.toml" is because we still use
"setup.py" for packaging and `pip` would get confused by the presence of
"pyproject.toml". The `[build-system]` section just tells it to look for
"setup.py".

[0]: https://black.readthedocs.io/en/stable/version_control_integration.html#version-control-integration