Revert compliance with codacy complaint
This reverts commit 27c785cb05f2f365994e32bfeaa0f31d7a19651e. While at
it, I also found some places, where I thought it beneficial to reword
the docstring. I'm putting these changes into this commit because the
difference is easy to spot and that way somebody going through the
commit history doesn't have to read the same similar line twice.
Replacing "data set" with "dataset" was also done deeper in the
docstring so that the wording stays consistent.
Originally I only wanted to use the revert to get an easy commit so that
I can test out whether the new Travis setup works as expected.