Deduplicate metadata "licenses" key
If I'm correct both "licenses" keys on the different dictionary levels
point to the same value. I extracted the value, put it into a local
variable and used that one instead. This removes code duplication, makes
the value more readable, as there is more horizontal space to write
it out, and it also makes the dictionary a bit more readable as it
removes some of the nesting.