Wrap multi line strings in parentheses
At least IMHO this improves the readability of the metadata dictionary.
For me it was pretty hard to spot that there were dictionary keys
immediately followed by strings spanning multiple source code lines.
Wrapping the values in parentheses makes it a lot easier to spot when
the value starts, when the next dictionary entry starts and it also
makes the multi line string more readable. At least for me, that is.
I put the space separating words at the start of each line, except the
first, which doesn't need one. This makes it a lot easier to spot
missing spaces, which is evidenced by the fact that I found a missing
space between "using" and "osm2pgsql" that way.
Last but not least I changed 'reference date' to '"referenceDate"'
because that's the way the key is spelled in the metadata dictionary.