Remove a redundant module layer
The `import_osm` under `egon.data.importing.openstreetmap` was a bit
redundant, because all the information was already present in the naming
of the parent modules. While having code in the `__init__.py` file is a
rarely used feature, we have a legitimate reason to do so here, because
we need the "openstreetmap" folder to contain the style file. This means
we can not have an "openstreetmap.py" file, but we're forced to use the
"__init__.py" file in the directory.
If we ever identify further modules which should be present below
`importing.openstreetmap` and which can be used to structure the code in
a better manner, we can freely move the functions around and re-export
them from "openstreetmap/__init__.py" so that other imports don't have
to be changed, a behaviour which is also documented in the module's
docstring.