Add the API reference rST files
These where generated via:

```
sphinx-apidoc --implicit-namespaces --no-toc -o docs/reference \
              src/egon
```

This has to be done every time a new source file is introduced. Maybe we
can automate it. We'll see.

The `--no-toc` flag is necessary, because the `index.rst` file present
in the directory already fulfills that purpose.
The `--implicit-namespaces` flag is necessary because `src/egon` doesn't
(and can never) contain an `__init__.py` file, so without the flag,
Sphinx wouldn't recognize `oemof` to be the root package's name.