Make `egon.data.processing` an explicit package
This fixes the tests failing with an error because they're importing two
different `openstreetmap` modules from two different locations but
getting confused between the two. There's an [issue comment][0] and a
[documentation section][1] explaining this. Unfortunately adding the
`--import-mode=importlib` option didn't fix the error so I had to use
the `__init__.py` fix. Maybe it's a doctest specific thing.

[0]: https://github.com/pytest-dev/pytest/issues/3151#issuecomment-360493948
[1]: https://docs.pytest.org/en/latest/pythonpath.html#import-modes