Switch DAGs "On" by default
The on/off button toggling the pause state is easy to miss which leads
to a lot of confusion and even bug reports. This commit changes the
configuration to switch DAGs "On" by default, both on a global level, as
well as on the individual "egon-data-processing-pipeline". The latter is
redundant, but kept in the code as an example, because the parameter was
surprisingly hard to find.
This fixes #33 and ticks of an item in issue #65.