Don't run the pipeline on a regular schedule
Not specifying a `schedule_interval` means that Airflow assumes a daily
schedule. This means that if the scheduler sees that a DAG is not paused
and behind schedule, the DAG will be started automatically. Since we
only want to start our DAG manually, we have to set `schedule_interval`
to `None`.