Configure revision script filenames
Just using the revision id doesn't convey too much information about the
script when looking at the filename. It especially doesn't convey any
information with respect to the relative ordering among scripts. Adding
the time stamp to the script's filename at least gives an idea about
this and it also gives a hint on how recent the change is. One even
might use this information to figure how whether old migrations should
be pruned in the future. When one talks about timestamps, it's of course
good to always have them in UTC. That way, they are comparable even
without specifying time zone information, which would just add
unnecessary noise here.
I also changed the character separating the different parts of the
filename from underscores to periods, because those are already used to
separate the suffix from the rest of the filename and using a smaller
number if different characters is easier on the eyes, IMHO.
Last but not least, I relaxed the length at which the revision message
to 53 characters. This is the maximum length of the first line of a Git
commit message, i.e. 50 characters, with a little bit of leeway.