Prefix migration scripts with a sequential number
Since this moves the migration script, this hook has to be run after
every other hook. Everything else should be explained in the comments
and docstrings.
We also have to set `revision_environment` to `true` because the
"env.py" file contains our hook definition which always has to be loaded
in order to move a generated migration script.