Add the base migration
The file was generated using the command

```
alembic revision -m "Do nothing for the base migration"
```

inside the "src/egon/data/alembic" directory. This is important, as this
directory contains the "alembic.ini" file which Alembic needs to
operate.

As the docstring states, this migration is initially empty in order to
provide a blank slate to build upon and to go back to.