Add a migration creating database extensions
The extensions in question are hstore and postgis. These are currently
added via a Docker entry point. Since this already created problems and
since it also wouldn't work when not using Docker, moving it into a
migration is both, a good proof of concept and a fix for an open issue.