Generate Airflow's/Flask's secret key randomly
Airflow version 1.10.14 changed its behaviour to disallow starting up
with `secret_key` set to the insecure value of `temporary_key` and
changed the configuration default to `{SECRET_KEY}` which is a special
value allowing Airflow to generate a random key internally. This commit
just adopts this new configuration default, allowing us to work with
Airflow versions >=1.10.14 again, at least when it comes to this issue.
For more information, look into the corresponding [Airflow commit][0]
and Airflow's [CHANGELOG][1]. Search for "#3738" and "#12747" to find
the entries of interest.
This fixes #83.

[0]: https://github.com/apache/airflow/pull/3738/commits/9e01476a50b9be27c4b1e6c6e24d36f290629195
[1]: https://github.com/apache/airflow/blob/ede67753f18dc020c5081b03b6bb4cfadecb34da/CHANGELOG.txt