Replace non-toplevel usages of `subprocess.run`
Using `subprocess.run` at the top level of the CLI, where it basically
replaces the main process, is OK because exceptions terminate the
program and the stack traces are immediately visible. But in airflow
tasks, `subprocess.run` should be replaced by `egon.data.subprocess.run`
because it makes tasks fail on errors and the error messages get written
to the logs.