Fix importing the zensus population data
The problem was that the file was extracted relative to the current
working directory, but the path to the file that was going to be
imported was constructed as pointing into the location of the
`egon.data.importing.zensus` package. When the `__init__.py` file is run
directly from the directory it is contained in, the generated location
coincides with the current working working directory. But in general,
these two locations will be different.