Manually break lines to be less than 80 characters
The first line was pretty interesting because it contained exactly 80
characters. Other than that, breaking things up was pretty simple and
there are several degrees of freedom in how can distribute the strings.
The most important thing to know is that consecutive lines consisting
just of strings are automatically concatenated by the Python
interpreter.