Use a raw string because it contains a backslash
Otherwise one of the style checkers complains because of an invalid
escape sequence and the Python interpreter might to weird stuff if the
backslash accidentally leads to a valid escape sequence. So if you want
to include as verbatim backslash in a string, it's best to use a raw
string.