Change the error checking `docker-compose` command
The proposed command only works in a directory containing a
"docker-compose.yaml" (or .yml) file. The new one supplies the file
inline, so it works everywhere. Also, using the `ps` subcommand instead
of `up -d --build` means less confusing command line switches and in
case of success, the subcommand essentially does nothing, so it's done
faster and it has less, potentially confusing, output.