Differentiate the error types
There are "Installation" and "Runtime Errors" and after this commit the
"Troubleshooting" section differentiates between them by putting them in
matching subsections. To this end, the error sections have also been
reordered, because the runtime error came before the installation
errors. The catch-all section is the last one and is one level higher,
because it might refer to both, installation or runtime errors.

Since the two error sections are now referred to separately in the
README, the state of the "Further Reading" section has been reverted
back to its state before commit
58dc0f64171e51b9f9268cdb2758173dae1b33c3.