Change "issued-id" to "issue-id"
This could have been a typo, or it could mean that what is meant is the
id that was "issued" by GitHub. Just using issue avoids this confusion,
because now it's clear that the placeholder stands for the id of the
issue. In order to be really sure though, I've also added a short
explanatory sentence, driving home the point, that this is just a
placeholder.