Adjust footnote block indentation
Switching to automatically numbered footnotes means we have identifiers
of varying length again. In the spirit of not having indentation depth
be dictated by identifier length, I put started the footnote block on a
line by it's own so that it an simply be indented by the structurally
appropriate amount.
Otherwise, the text wasn't changed. (I think.)