Don't use `git commit -m` in the example
The `-m` switch promotes the exception, which is having a commit which
can be summarized by a single line, to the default. Instead, most of the
times it's useful to have a longer commit message explaining the changes
in more detail. Also, the example commit message did not adhere to the
general style guide by including a dot at the end.