Read SQL scripts via `importlib_resources.read_text`
That's the easiest way to make sure that the path to the SQL scripts is
always tied to the path of the package they are located in. The downside
is that we have to install the external "importlib-resources" package,
until we drop support for Python versions lower than 3.9, after which we
can move from `importlib_resources` to `importlib.resources`.