Wrap main CLI command in an initialization function
This makes sure that the module doesn't contain executable code besides
the function definition. The CLI entry point is still `main` but now
`main` does all setup and initialization and then invokes the main CLI
command. This is also the only way of catching exceptions raised during
the commands execution, which has to be used later on.