Support proper configuration files
If the configuration file doesn't exist, create one on startup by
writing the default values of all command line options having defaults
to a file in the correct format. In order for this to work, the command
line options actually need defaults, so specify those and make them be
shown in the `--help` message. Also put a note in the help message,
explaining how to configure `egon-data` as well as the fact, that it
will modify it's working directory. Since `egon-data` might be run
multiple times in parallel, also write the combination of configuration
values and command line overrides to a special configuration file from
the first `egon-data` process started. Exclusively use the options from
this configuration file in all other processes and make sure that the
configuration file is deleted when the process exits by wrapping the
main command in a `try` block where the file containing the running
processes active option configuration is deleted in a `finally` block.

Put a configuration file path discovery function in the
`egon.data.config` module in order to support all this machinery.

Finally, adapt the CHANGELOG.rst to these new changes.