Make "Docker"ed database configurable
Allow configuring the "Docker"ed database via command line arguments.
In order to do so, convert the shipped "docker-compose.yml" to a
template, which is copied to the "docker" subdirectory of the current
working directory, with the template fields filled in with the
corresponding values from the command line arguments. Run
`docker-compose` using the copied template, instead of the original file
shipped with "egon.data". Put a note about this behavior into
`egon-data`'s help message. Stop reading database parameters from the
shipped `docker-compose.yaml` because that's no longer possible, as the
file is a template without meaningful values now. Finally, update the
changelog to reflect this new feature.