Blacken "src/egon/data/db.py"
I intentionally violated the style in the last commit, in order to keep
the diff as small as possible. While fixing the style, I also changed
the variable `docker_db_config` to `configuration`. Using the same
variable name that is also used to store the "egon-data" part of the
configuration file is done intentionally, because the new
`configuration` dictionary is just the old one, with the keys filtered
to the ones also present in `translated` and translated accordingly.