Use the "Docker"ed database for Airflow
This has a lot of consequences. First and foremost, the
"sql_alchemy_conn" configuration setting in "airflow.cfg" has to be
determined dynamically from the database configuration settings. In
order to do so, "airflow.cfg" was converted to a template, where
everything in curly brackets, e.g. "{slot}", gets filled with
appropriate values, which also means that curly braces not signifying a
slot to be filled have to be escaped to "{{" and "}}". This is the
cause for most of the changes in "airflow.cfg".

The template gets handled similar to "docker-compose.yml", in that it
gets copied to the "airflow" subdirectory of the current working
directory and has its slots filled. As a consequence, the directory in
which DAGs are discovered can no longer be determined from
`$AIRFLOW_HOME` but has to be filled in as well, because our DAGs still
reside in the "egon.data" package. This explains having to set the
"dags_folder" configuration setting.

Since the one existing database has a new purpose, it also gets a
setting, with which it can be configured and which is used in
"docker-compose.yml" to determine the database name. The database now
has to be up prior to Airflow, so we can no longer initialize it in an
Airflow task. Therefore starting the docker container is now done at the
start of the `egon-data` command. Since the docker container can only be
configured to contain one database initially, it's now checked whether
the normal `egon-data` database exists and if it doesn't it is created.
That way one can use multiple different databases, simply by supplying
different database names via `--database-name`.

Last but not least, the "entrypoints" make no sense anymore, because
those would be applied to the Airflow metadata database. Extension
creation and other initialization has therefore been moved to the
`initdb` Airflow task, which finally fixes #76.