Fix combining configuration settings
The original implementation using `dict(d1, **d2)` only did a shallow
copy, which defied it's purpose, because the top level of the
configuration settings always contained the command name `"egon-data"`.
This meant that the new configuration settings for this command simply
overwrote the old ones, instead of doing a proper merge. Now there's a
proper `merge` implementation doing a recursive merge, so that no
configuration settings are lost anymore.