Write unknown settings to configuration files
If a configuration file was present but did not contain a setting for a
certain configuration option, the existing file was never updated to
contain the new option. Now newly implemented options are automatically
written to a configuration file when they are first used.
Note the somewhat reverse order in the `merge` call. This is done in
order to overwrite values supplied on the command line with the ones
already present in the configuration file, so that these ones don't get
silently changed. If a value is not present in the `stored` command line
options, it will retain it's value from the new options which will
subsequently be stored in the new configuration file.