Disambiguate and streamline the names in the compose file
Before this commit, there was an ambiguity in the name
`egon-data-local-database`: the name was used for both, the service as
well as the container. As the container name will get configurable, it's
better to use non ambiguous names from now on. While doing this, I also
noticed that the name `egon-data-database-volume` does not use the same
convention as the two new names introduced. In order to be consistent
when naming things in this file, I also changed the name to
`egon-data-database-volume`.