Store the "egon_data" connection in Airflow's DB
The code for this is more complex than using environment variables, but
it persists even after the main process stopped which makes it easier to
debug. It also ensures that the same connection is shared across all
Airflow instances using the same metadata database. Last but not least,
using Airflow's metadata database to store values opens up an
alternative way to store the current `egon-data` configuration so this
is also a test for future changes.