Outdent assignment block
The indentation isn't really necessary. It's only there because if the
`custom.is_file()` files, there is no `configuration` to translate, so
the assignment would fail with a `NameError` or a similar exception.
Therefore the assignment block was placed inside the `if` block. But the
return would fail for the same reason, so the assignment can be placed
outside of the `if` block.