Adjust tests to standalone `egon_data` function
I didn't adjust the tests when separating the `egon-data` logic from
`main` which obviously made the tests fail. Now the tests no longer
invoke `main`, which no longer is a `command`, but they invoke
`egon_data`, which is the new name of the main command containing the
`egon-data` logic, so now the tests should pass again.
This actually fixes #183.