Always initialize Airflow's DB on startup
Previously, the Airflow's metadata database was only initialized when
starting the webinterface. But since we now store connection information
in the metadata database on startup, we need the database initialized,
so it's now done every time.
While at it, I also used the new Airflow 2.0 way of initializing the
database, so there's one less thing to worry about when migrating.