Explicitly set `Connection` attributes
Since `uri` is only a constructor parameter and not actually an
attribute of `Connection` objects, setting it doesn't seem to do
anything, which might be the reason for `PostgresOperator`s not picking
up connection parameters, leading to #193.
Note that `Connection`s don't have an attribute for the database name,
but from testing it out, it seems that the database name is stored in
the `schema` attribute.