Move "Tips" to "CONTRIBUTING.rst"'s bottom
This is just a rather unimportant leftover of the automatically
generated project skeleton. The section isn't very informative and needs
to be fleshed out quite a bit more, so placing it at the bottom is
probably the safest bet for now. While at it, I also lifted the
heading one level upwards to reduce unnecessary nesting.