Remove "But reports" and "Feature requests..."
The content in these sections was also duplicated at a different
location within the file. Not the content appears only once. While at
it, I also rewrote and reformatted the first paragraph plus the
following line in an unconventional way in order to keep the diffs small
and make the upcoming merge a bit easier.