Fix assignment of next subtstation
The assignment of substations to next voronoi cut that includes a
substation still has one wrong assignment. It can happen that parts of
the municipality x voronoi cuts are adjacent to an already assigned part
but are assigned to a different substation (without touching any of the
polygons that are assigned to this susbtation).