Place OSM files below the working directory
This mostly amounts to changing the location of the target file from
`Path(__file__).parent` to `Path(".") / "openstreetmap"`. But there are
two things which have to be taken special care of. The first one would
be to ensure that the new location actually exists, which is done by
creating it after checking whether the file already exists. The second
important thing is that the input file is now supplied via an absolute
path via `input_file.absolute()` because `osm2pgsql` is executed in a
subdirectory of `egon-data`'s installation directory by setting
`cwd=Path(__file__).parent`.