Replace `__file__` with `importlib_resources.files`
It's safer, because that way the actual alias of the imported module is
used, meaning that there is less opportunity for typos. Also, if the
module is moved, only the import has to be adjusted. This fails earlier
than the dynamic resolution of a relative path, so it's easier to debug.
The code is also more readable now (IMHO), and we want to replace usages
of `__file__` or `__package__` with the more stable
`importlib_resources` anyway.