Store database data in a `bind mount` again
This undoes 90f281179dac7deeb05ab171a381b7feab03323a because apparently
its hard to change the storage location of Docker volumes on a per
volume basis. One of the things that changed since the commit mentioned
above though, is that the pipeline is now supposed to be run in a
dedicated working directory and that the "docker-compose.yml" file from
the repository is used as a template to generate a compose file specific
to this working directory. This means that we no longer have to put the
database data into a central location, but we can put it relative to the
location of the generated "docker-compose.yml" file.