Use host's numeric ids inside the container
It seems that when using bind mounts, Docker changes the host's owner of
the folder mounted into the container. This may create a folder
owned by a user which doesn't exist on the host. If this happens, and
the container is stopped and restarted, the container fails to mount
the folder because it can't access it. In order to prevent this, the
container has to be set to run its internal processes under the same
numeric user and group id under which the mounted folder should be owned
on the host. This commit does so by using the current processes user and
group ids.