Replace ad-hoc dictionary with proper dataclass
Its a class collecting certain information about a `TaskGraph`, namely
the first, i.e. initial tasks, the last, i.e. final tasks and the
complete collection of tasks in a `TaskGraph`. Using a proper class
instead of a dictionary for this opens up a space where documentation
can later be placed and makes the API less arbitrary.