Unconditionally skip `Dataset` tasks
The way this works is that for each task that is part of a `Dataset`, a
`class` is dynamically created as a subclass of the tasks `__class__`.
The new class is then set to be the `__class__` attribute of the task,
which inserts it into the inheritance hierarchy of this particular task.
As the new class overrides `execute` to unconditionally raise an
`AirflowSkipException`, the task is always skipped.
This is a proof of concept showing that skipping a task this way is
possible, which is why tasks are skipped unconditionally for now. The
implementation of `execute` is generated by a method of the `Dataset`
instance so that `execute` has access to the `Dataset`'s version, which
can later be used to conditionally skip the task. This will be
implemented in a later commit though.