Allow using the `Dataset` class as a dictionary
That way it can be used to store all datasets keyed by name.

For those who care about implementation details: normally, the metaclass
mixin could have been implemented as

```python
class UserDictMetaclassMixin(type, UserDict):
    pass
```

but that lead to the following troubles:

  - the implementation did not automatically create a class level `data`
    attribute and
  - the implementation did not work in conjunction with the `@dataclass`
    decorator.

The first problem could be remedied by overriding `__init__` to add the
`data` attribute. The second problem was caused by the fact that under
certain circumstances `@dataclass` uses the decorated class' `__eq__`
implementation, expecting the equality testing semantics of classes and
arbitrary objects, which usually defers to the semantics of `is`. Using
the above mixin implementation, equality testing would use the
implementation inherited from `UserDict.__eq__` which tests for equality
with the `data` dictionary. This is not what we want for the class. The
solution is to supply a custom `__eq__` method which is based on `is`.
Testing for equality with the dictionary can always be done using the
`data` class attribute.