Revert "dictionarizing" the `Dataset` class
This reverts commit 46556d927e30db718f30d2ac1f42fb3b1dabe5fc because
there are still some errors with having a DAG whose defining module
imports this implementation of being able to use classes (NOT their
instances) as dictionaries. In particular, it seems that Airflow tries
to serialize and restore these classes while loading the DAG, which
triggers an error that is hard to debug.
Because this error is rather hard to track down, the implementation is
removed for now, until there's more time check what's wrong with it and
which interactions trigger the error.