Convert module constant `dataset` to function
Having this as a module constant means that there is only one `dataset`
instantiated for the lifetime of a Python process, which in turn means,
that the tasks are only instantiated once, too. Airflow seems to
instantiate the DAG more than once which is a problem, because a task
can only be assigned one DAG. Afterwards, the task can not be reassigned
to a different DAG. This means that for each new DAG Airflow creates,
new tasks have to be created, which means that a new `Dataset` instance
has to be created. Having `dataset` be a module constant prevents
exactly this, which is fixed by having it be a function which is
re-evaluated whenever a DAG is created.
The formatting intentionally defies the style guidelines in order to
keep the diff minimal and not obscure the actually important changes
with lines only changing because of a change in indentation level.