Attach a condition to skipping `Dataset` tasks
The condition/implementation should be fairly obvious: check whether the
tasks dataset is already marked as executed for the currently specified
version. If so, skip execution. If not, execute the task.
Note that since no execution information is actually persisted to the
database for now, the tasks of every `Dataset` will always be executed
until writing/deleting version information to/from the database is
implemented.