Remove a `Dataset`s version information on updates
This is important in order to mark all dependent `Dataset`s for
re-execution, which is [one of the reasons][^0] that `cascade="delete"`
is added to `Dataset` `Model`s `dependents` relationship. Specifying
`cascade="delete"` here has the effect of deleting all dependent
`Dataset`s if a dependency gets deleted. So if a `Dataset`'s version
changes, its tasks are re-executed and its version information is
deleted from the database, which triggers the deletion of all version
information of its dependents. This in turn means that all tasks of
dependent `Dataset`s also get re-executed, even if their version is
unchanged.
Note again, that all of this currently doesn't really have an effect
since no version information is stored in the database for now. Tasks
are always executed and the loop will always cover an empty
collection, never executing its body. Persisting version information to
the database will be implemented in the next commit though.

[^0]: Also, SQL usually refuses deleting objects if related objects
      aren't deleted too, so referential integrity forces us to do this
      too, but I'd rather explain the rationale for setting the database
      up this way.