Persist `Dataset` version updates to the database
Doing this right is a bit complicated, which is why the commit is rather
large..

Actually storing a `Dataset`'s version and dependency information in the
database is done in the `update` method of a `Dataset`. Since this
method needs to modify the data stored in the database, it obviously
needs a `session` to work with. The method is either called as part of
the last task of a `Dataset` in which case the method shares the session
scope in which the task is executed. This is done by adding an
`after_execution` parameter to the `check_version` wrapper. This
parameter is a sequence of functions that are to be run after the
version checked task has been executed. The sequence is empty for all
tasks of a `Dataset` except the last task. For the last task the
sequence only holds the `update` function which thus gets executed after
all task got successfully executed.

Now this only works if a `Dataset` only has one last task. If there are
multiple tasks which are to be executed in parallel at the end of a
`Dataset`, it's impossible to know in advance which one will actually be
the last one to be finished. For this reason, an explicit final task
only updating version and dependency information needs to be created for
this case, which is done by explicitly testing for it via:

```
if len(self.tasks.last) > 1:
```

The if block then establishes a single final `Dataset` task regardless
of how the `Dataset` was previously structured. This task is then
selected, in order to add updating version information to it when
modifying the task to be version checked.