Increase the indentation level of `connect`'s body
There are no changes in the behaviour of the code here. This is purely
done because I'm going to convert `connect` into the `__init__` method
of `Tasks`, for which the indentation level would have to be increased,
too. In order to not obscure the actually significant behavioural
changes, I'm separating the increase in indentation level from the other
changes with this commit.