Store the initializer argument in `Tasks.graph`
This fixes a bug with adding the `update_version` task to a `Dataset`,
because `(tasks, update_version)` wasn't a valid `TaskGraph`. If
`tasks.graph` is a valid `TaskGraph`, then so is
`(tasks.graph, update_version)` but this is only available now because
now we save the actual `TaskGraph` used to initialize a `Tasks`
instance. This bug fix is the last line in this commit.
The initializer argument is now also named `graph` which is more
sensible given that it's the `TaskGraph` with which the `Tasks` instance
is initialized.