Make `Tasks` a `dict` subclass
A `Tasks` instance now also acts as a dictionary with `task_id`s as keys
and the actual tasks as values. This obviates the need for the `all`
attribute, because one can now simply access all tasks via the `.values`
method.
What could be deemed a little complicated here is the line

```
self.update(reduce(lambda d1, d2: dict(d1, **d2), results, {}))
```

but this is just a compact way of merging multiple dictionaries into one
and updating the current `Tasks` instance with the merge result.