Rename classes generated to check versions, again
Since the name of a class generated dynamically via `type` can contain
spaces, there's no need to adhere to the camel case naming convention
for those classes. Therefore, the "versioned" annotation can actually be
placed in parenthesis and separated with a place after the class name,
and the dataset name from which the class name is generated can also
contain a space and written more like regular text.