Allow unversioned tasks as `Dataset` dependencies
Such tasks can't be used to check whether a downstream task should be
rerun because they don't have a `Dataset` associated with them. So these
tasks have to be filtered out when storing the `Dataset`s whose update
causes a rerun of a particular downstream task. In order to do this, all
that has to be done is check whether the `.dataset` attribute is present
on a task and not try to store its `Dataset` as a dependency, if it
isn't.