Add random seed to CLI parameters
Some tasks use python's native or numpy's random package. To make sure
to get the same results for multiple runs of the pipeline (or at least
to rule out this source of error), the drawings are to be in same order.
Therefore, the same seed needs to be used on init.

You can obtain the seed in your tasks by
`SEED = config.settings()['egon-data']['--random-seed']`
Subsequently, set the seed by `random.seed(SEED)` or
`numpy.random.seed(SEED)` respectively.

Pay special attention to tasks which make use of multiprocessing!
(cf. #351)