oss arguments to subprocess.run as a sequence
This is the recommended way, unless `shell=True` is set. This commit
also removes `shell=True` because it adds unnecessary system dependent
behaviour and allows us to go back to the recommended way of passing
arguments as a sequence.
Note that the argument sequence had to be modified because it wasn't set
up correctly. More specifically, some strings which should've been
separate entries in the sequence where already `" ".join`ed, so I had to
separate them.