Run osm2pgsql in the current working directory
In order to do this, we have to tell `osm2pgsql` where to find its style
file. This is packaged with `egon-data`, so we can get the path to the
style file via `importlib_resources.path`, but only as a context
manager. While in theory, one could put generating the command sequence
and running it in the context manager's block, that would introduce a
lot of additional indentation, making the diff unnecessarily big. So
instead, the path gotten from the context manager is used to copy the
style file from `egon-data`'s installation location to a subdirectory of
the current working directory. After this, the style file is in a known
location which can be passed to `osm2pgsql`, removing the need to set
the `cwd` of `osm2pgsql` and instead allowing it to run in the current
working directory which all other Airflow tasks should use too.