Only store Airflow operators in `Tasks.graph`
While traversing the constructor argument of `Tasks`, leafs which aren't
already Airflow operators are converted to operators. This leads to the
problem that when trying to create `Tasks` instance from another
`Taks`' `graph`, Airflow operators are created and added to the Airflow
DAG, which are already present, which is currently ignored with a
warning by Airflow and it will be an error in Aiflow 2.0 and later.
Currently, this means that changes to the newly created operators are
not reflected in the DAG, because the new operators are never added to
the DAG.
With this commit, instead of storing the original arguments in the leaf
of a `Tasks.graph`, rather the converted operators are which means that
operators are created only once, so any changes to them are actually
reflected in the DAG they are present in.
This fixes #365.