Remove unknown interpreted text roles in README.rst
The used roles are only known by Sphinx. But since README.rst is also
displayed on GitHub which only parses vanilla reST, these roles can't be
used here.
The `:py:mod:` role can just be replaced by `:code:` but in order to
replace the `:ref:` role the link target had to be adjusted.